feat: craate getAuthClient helper function and use in oauth/invoices example
