+++ examples/.gitignore
@@ -1,3 +1,4 @@
 node_modules
 .yarn/install-state.gz
 .pnp.*
+.env
