+++ examples/oauth/.env.example
@@ -1,0 +1,6 @@
+# OAuth Clients
+CLIENT_ID=""
+CLIENT_SECRET=""
+
+# Access Token
+ACCESS_TOKEN=""
