+++ examples/oauth/invoices.ts
@@ -1,43 +1,16 @@
-import * as readline from "node:readline/promises";
-import { stdin as input, stdout as output } from "node:process";
+import { Client, GetInvoicesRequestParams } from "@getalby/sdk/oauth";
+import { getAuthClient } from "./helper";
 
-import {  Client, GetInvoicesRequestParams, OAuth2User } from "@getalby/sdk/oauth";
+const userAgent = "AlbySDK-Example/0.1 (invoices-demo)";
 
-const rl = readline.createInterface({ input, output });
+const auth = await getAuthClient(userAgent);
 
-if (!process.env.CLIENT_ID || !process.env.CLIENT_SECRET) {
-  throw new Error("Please set CLIENT_ID and CLIENT_SECRET");
-}
+const client = new Client(auth);
 
-
-const authClient = new OAuth2User({
-  client_id: process.env.CLIENT_ID,
-  client_secret: process.env.CLIENT_SECRET,
-  callback: "http://localhost:8080",
-  scopes: ["invoices:read", "account:read", "balance:read"],
-  user_agent:"AlbySDK-Example/0.1 (invoices-demo)",
-  token: {
-    access_token: undefined,
-    refresh_token: undefined,
-    expires_at: undefined,
-  }, // initialize with existing token
-});
-
-console.log(`Open the following URL and authenticate the app:`);
-console.log(await authClient.generateAuthURL());
-console.log("----\n");
-
-const code = await rl.question("Code: (localhost:8080?code=[THIS CODE]: ");
-rl.close();
-
-await authClient.requestAccessToken(code);
-console.log(authClient.token);
-const client = new Client(authClient);
-
-const params:GetInvoicesRequestParams = {
+const params: GetInvoicesRequestParams = {
   page: 1,
-  items: 5
-}
+  items: 5,
+};
 const response = await client.incomingInvoices(params);
 
 console.log(JSON.stringify(response, null, 2));
