+++ examples/package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@getalby/sdk": "portal:../",
+    "dotenv": "^17.2.2",
     "websocket-polyfill": "^0.0.3"
   },
   "devDependencies": {
