+++ package.json
@@ -52,7 +52,7 @@
     "lint:js:fix": "eslint src --ext .js,.ts --fix",
     "tsc:compile": "tsc --noEmit",
     "format": "prettier --check '**/*.(md|json)' 'src/**/*.(js|ts)' 'examples/**/*.(js|jsx)'",
-    "format:fix": "prettier --loglevel silent --write '**/*.(md|json)' 'src/**/*.(js|ts)' 'examples/**/*.(js|jsx)'",
+    "format:fix": "prettier --loglevel silent --write '**/*.(md|json)' 'src/**/*.(js|ts)' 'examples/**/*.(js|jsx|ts)'",
     "prepack": "yarn run build",
     "test": "jest",
     "clean": "rm -rf dist",
